+++ HW09-JDBC/src/main/java/ru/otus/businessLayer/service/DBService.java
@@ -1,0 +1,12 @@
+package ru.otus.businessLayer.service;
+
+import java.util.Optional;
+
+public interface DBService<T> {
+
+    Object saveObject(T object);
+
+    Object saveObject(T object, boolean flagOfInsert);
+
+    Optional<T> getObject(Object id);
+}
