+++ HW09-JDBC/src/main/java/ru/otus/businessLayer/service/DbServiceException.java
@@ -1,0 +1,7 @@
+package ru.otus.businessLayer.service;
+
+public class DbServiceException extends RuntimeException {
+    public DbServiceException(Exception e) {
+        super(e);
+    }
+}
