--- HW09-JDBC/src/main/java/ru/otus/core/annotations/Id.java
@@ -1,11 +1,0 @@
-package ru.otus.core.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Target(ElementType.FIELD)
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Id {
-}
