--- HW09-JDBC/src/main/java/ru/otus/core/dao/ObjectDaoException.java
@@ -1,7 +1,0 @@
-package ru.otus.core.dao;
-
-public class ObjectDaoException extends RuntimeException {
-    public ObjectDaoException(Exception ex) {
-        super(ex);
-    }
-}
