--- HW09-JDBC/src/main/java/ru/otus/core/service/DBService.java
@@ -1,10 +1,0 @@
-package ru.otus.core.service;
-
-import java.util.Optional;
-
-public interface DBService<T> {
-
-    int saveObject(T object);
-
-    Optional<T> getObject(int id);
-}
