--- HW09-JDBC/src/main/java/ru/otus/core/service/DbServiceException.java
@@ -1,7 +1,0 @@
-package ru.otus.core.service;
-
-public class DbServiceException extends RuntimeException {
-    public DbServiceException(Exception e) {
-        super(e);
-    }
-}
