--- HW09-JDBC/src/main/java/ru/otus/core/sessionmanager/DatabaseSession.java
@@ -1,4 +1,0 @@
-package ru.otus.core.sessionmanager;
-
-public interface DatabaseSession {
-}
