--- HW09-JDBC/src/main/java/ru/otus/core/sessionmanager/SessionManager.java
@@ -1,13 +1,0 @@
-package ru.otus.core.sessionmanager;
-
-public interface SessionManager extends AutoCloseable {
-    void beginSession();
-
-    void commitSession();
-
-    void rollbackSession();
-
-    void close();
-
-    DatabaseSession getCurrentSession();
-}
