--- HW09-JDBC/src/main/java/ru/otus/core/sessionmanager/SessionManagerException.java
@@ -1,12 +1,0 @@
-package ru.otus.core.sessionmanager;
-
-
-public class SessionManagerException extends RuntimeException {
-    public SessionManagerException(String msg) {
-        super(msg);
-    }
-
-    public SessionManagerException(Exception ex) {
-        super(ex);
-    }
-}
