+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/core/annotations/Id.java
@@ -1,0 +1,11 @@
+package ru.otus.daoLayer.core.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.FIELD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface Id {
+}
