+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/core/dao/ObjectDao.java
@@ -1,0 +1,14 @@
+package ru.otus.daoLayer.core.dao;
+
+import ru.otus.daoLayer.core.sessionmanager.SessionManager;
+
+import java.util.Optional;
+
+public interface ObjectDao<T> {
+
+    Object insertObject(T object, boolean flagOfInsert);
+
+    Optional<T> findById(Object id);
+
+    SessionManager getSessionManager();
+}
