+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/core/dao/ObjectDaoException.java
@@ -1,0 +1,7 @@
+package ru.otus.daoLayer.core.dao;
+
+public class ObjectDaoException extends RuntimeException {
+    public ObjectDaoException(Exception ex) {
+        super(ex);
+    }
+}
