+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/core/sessionmanager/DatabaseSession.java
@@ -1,0 +1,4 @@
+package ru.otus.daoLayer.core.sessionmanager;
+
+public interface DatabaseSession {
+}
