+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/core/sessionmanager/SessionManager.java
@@ -1,0 +1,13 @@
+package ru.otus.daoLayer.core.sessionmanager;
+
+public interface SessionManager extends AutoCloseable {
+    void beginSession();
+
+    void commitSession();
+
+    void rollbackSession();
+
+    void close();
+
+    DatabaseSession getCurrentSession();
+}
