+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/core/sessionmanager/SessionManagerException.java
@@ -1,0 +1,12 @@
+package ru.otus.daoLayer.core.sessionmanager;
+
+
+public class SessionManagerException extends RuntimeException {
+    public SessionManagerException(String msg) {
+        super(msg);
+    }
+
+    public SessionManagerException(Exception ex) {
+        super(ex);
+    }
+}
