+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/mapper/EntityClassMetaData.java
@@ -1,0 +1,17 @@
+package ru.otus.daoLayer.mapper;
+
+import java.lang.reflect.Constructor;
+import java.lang.reflect.Field;
+import java.util.List;
+
+public interface EntityClassMetaData<T> {
+    String getName();
+
+    Constructor<T> getConstructor();
+
+    Field getIdField();
+
+    List<Field> getAllFields();
+
+    List<Field> getFieldsWithoutId();
+}
