+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/mapper/EntityClassMetaDataException.java
@@ -1,0 +1,7 @@
+package ru.otus.daoLayer.mapper;
+
+public class EntityClassMetaDataException extends RuntimeException {
+    public EntityClassMetaDataException(String s) {
+        super(s);
+    }
+}
