+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/mapper/EntitySQLMetaData.java
@@ -1,0 +1,12 @@
+package ru.otus.daoLayer.mapper;
+
+public interface EntitySQLMetaData {
+
+    String getSelectByIdSql();
+
+    String getInsertSql();
+
+    String getInsertWithoutIdSql();
+
+    String getUpdateSql();
+}
