+++ HW09-JDBC/src/main/java/ru/otus/daoLayer/mapper/JdbcMapperException.java
@@ -1,0 +1,11 @@
+package ru.otus.daoLayer.mapper;
+
+import java.util.function.Supplier;
+
+public class JdbcMapperException extends Exception implements Supplier<String> {
+
+    @Override
+    public String get() {
+        return "Object not created";
+    }
+}
