--- HW09-JDBC/src/main/java/ru/otus/jdbc/mapper/EntityClassMetaData.java
@@ -1,17 +1,0 @@
-package ru.otus.jdbc.mapper;
-
-import java.lang.reflect.Constructor;
-import java.lang.reflect.Field;
-import java.util.List;
-
-public interface EntityClassMetaData<T> {
-    String getName();
-
-    Constructor<T> getConstructor();
-
-    Field getIdField();
-
-    List<Field> getAllFields();
-
-    List<Field> getFieldsWithoutId();
-}
