--- HW09-JDBC/src/main/java/ru/otus/jdbc/mapper/EntityClassMetaDataException.java
@@ -1,7 +1,0 @@
-package ru.otus.jdbc.mapper;
-
-public class EntityClassMetaDataException extends RuntimeException {
-    public EntityClassMetaDataException(String s) {
-        super(s);
-    }
-}
