--- HW09-JDBC/src/main/java/ru/otus/jdbc/mapper/EntitySQLMetaData.java
@@ -1,11 +1,0 @@
-package ru.otus.jdbc.mapper;
-
-public interface EntitySQLMetaData {
-    String getSelectAllSql();
-
-    String getSelectByIdSql();
-
-    String getInsertSql();
-
-    String getUpdateSql();
-}
