--- HW09-JDBC/src/main/java/ru/otus/jdbc/sessionmanager/DatabaseSessionJdbc.java
@@ -1,17 +1,0 @@
-package ru.otus.jdbc.sessionmanager;
-
-import ru.otus.core.sessionmanager.DatabaseSession;
-
-import java.sql.Connection;
-
-public class DatabaseSessionJdbc implements DatabaseSession {
-    private final Connection connection;
-
-    DatabaseSessionJdbc(Connection connection) {
-        this.connection = connection;
-    }
-
-    public Connection getConnection() {
-        return connection;
-    }
-}
