+++ HW09-JDBC/src/main/java/ru/otus/jdbcImplementation/DbExecutorImpl.java
@@ -1,0 +1,43 @@
+package ru.otus.jdbcImplementation;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.sql.*;
+import java.util.List;
+import java.util.Optional;
+import java.util.function.Function;
+
+public class DbExecutorImpl<T> implements DbExecutor<T> {
+    private static final Logger logger = LoggerFactory.getLogger(DbExecutorImpl.class);
+
+    @Override
+    public Object executeInsert(Connection connection, String sql, List<Object> params) throws SQLException {
+        Savepoint savePoint = connection.setSavepoint("savePointName");
+        try (PreparedStatement pst = connection.prepareStatement(sql, Statement.RETURN_GENERATED_KEYS)) {
+            for (int idx = 0; idx < params.size(); idx++) {
+                pst.setObject(idx + 1, params.get(idx));
+            }
+            pst.executeUpdate();
+            try (ResultSet rs = pst.getGeneratedKeys()) {
+                boolean next = rs.next();
+                return rs.getObject(1);
+            }
+        } catch (SQLException ex) {
+            connection.rollback(savePoint);
+            logger.error(ex.getMessage(), ex);
+            throw ex;
+        }
+    }
+
+    @Override
+    public Optional<T> executeSelect(Connection connection, String sql, Object id, Function<ResultSet, T> rsHandler)
+            throws SQLException {
+        try (var pst = connection.prepareStatement(sql)) {
+            pst.setObject(1, id);
+            try (var rs = pst.executeQuery()) {
+                return Optional.ofNullable(rsHandler.apply(rs));
+            }
+        }
+    }
+}
