+++ HW09-JDBC/src/main/java/ru/otus/jdbcImplementation/sessionmanager/DatabaseSessionJdbc.java
@@ -1,0 +1,17 @@
+package ru.otus.jdbcImplementation.sessionmanager;
+
+import ru.otus.daoLayer.core.sessionmanager.DatabaseSession;
+
+import java.sql.Connection;
+
+public class DatabaseSessionJdbc implements DatabaseSession {
+    private final Connection connection;
+
+    DatabaseSessionJdbc(Connection connection) {
+        this.connection = connection;
+    }
+
+    public Connection getConnection() {
+        return connection;
+    }
+}
