--- HW09-JDBC/src/main/java/ru/otus/postgres/DataSourcePostgres.java
@@ -1,66 +1,0 @@
-package ru.otus.postgres;
-
-import javax.sql.DataSource;
-import java.io.PrintWriter;
-import java.sql.Connection;
-import java.sql.DriverManager;
-import java.sql.SQLException;
-import java.util.logging.Logger;
-
-/**
- * @author sergey
- * created on 03.02.19.
- */
-public class DataSourcePostgres implements DataSource {
-    private static final String URL = "jdbc:postgresql://localhost:5432/demoDB";
-    private static final String USER = "usr";
-    private static final String PASSWORD = "pwd";
-
-    @Override
-    public Connection getConnection() throws SQLException {
-        Connection connection = DriverManager.getConnection(URL, USER, PASSWORD);
-        connection.setAutoCommit(false);
-        return connection;
-    }
-
-    @Override
-    public Connection getConnection(String username, String password) throws SQLException {
-        return getConnection();
-    }
-
-    @Override
-    public PrintWriter getLogWriter() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void setLogWriter(PrintWriter out) {
-        throw new UnsupportedOperationException();
-
-    }
-
-    @Override
-    public int getLoginTimeout() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public void setLoginTimeout(int seconds) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public Logger getParentLogger() {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> iface) {
-        throw new UnsupportedOperationException();
-    }
-
-    @Override
-    public boolean isWrapperFor(Class<?> iface) {
-        throw new UnsupportedOperationException();
-    }
-}
