+++ build.gradle
@@ -8,7 +8,7 @@ plugins {
 
 idea {
     project {
-        languageLevel = 13
+        languageLevel = 15
     }
     module {
         downloadJavadoc = true
